repair tuple history when adding a column w/ a default value
The behavior of postgres when altering a table to add a default column
would lead to an inconsistency in our history capture:

- Subsequent query access sees column w/ default on existing rows
- No `ON UPDATE` trigger is fired for history-collection to observe this
- Live catalog queries do not match a query on latest snapshot

However, it is very typical that a client trying to extend a table
definition may wish to add a column with a default and with
`nullok=False`.  There is no other sequence of ERMrest requests which
can simulate this effect successfully.

So, as a workaround is to special-case this and make the implicit
row updates explicit during the POST column request:

1. Alter the table as before
2. Adjust the stateful model storage as before
3. If column.default_value is non-NULL, perform "redundant" update:
   - UPDATE tablename SET newcolumn = DEFAULT;

This way, the normal history-tracking trigger will fire and observe
the rows in a state consistent with what Postgres will present during
subsequent queries.