revise introspection-time safeguard to drop pg10 ermrest_history_* triggers
The master branch has a safeguard to drop ermrest_history triggers in
a scenario where a DBA creates a hazard by manipulating the DB schema
out of band.  Since that trigger is no longer relevant in this branch,
we need to revise the safeguard to delete the new statement-level
triggers which have the same hazard potential.