fix accidental clobbering of variable during model cache eviction
The model cache holds Python representations of catalog models keyed
by catalog and version metadata. It has an LRU policy to evict
older entries when it grows too large.

There was a mistaken reuse of a variable name within the eviction
loop which caused the final result of the catalog.get_model()
method to return an evicted entry rather than the intended live
entry on calls which had evictions.

This could exhibit as a range of malfunctions from a seemingly stale
policy or model validation check (if the eviction was a different
version of the same catalog's model) or seemingly random authorization
and model conflict errors if the eviction was for a completely
different catalog in a busy, multi-catalog server.