fix incomplete refactor for data interface which broke some authz logic
The previous refactoring patches had some inconsistencies in the handling
of the defaults query parameter for entity POST requests.  The fixes
touch quite a few of the refactored helper functions where we need to
consider the extra influence of the "use_defaults" plumbing concept,
which was a bit confused, mixing columns and column names.

Also fixed a broken entity POST test case that should have exposed
this problem and added a few related static authz test scenarios.