fix data-loss bug in history truncation
The history truncation feature could discard history and corrupt the
history-tracking facilities for tables whose lifetime straddles the
truncation horizon.  It was intended to prune history storage for
tables whose lifecycle finished prior to the truncation horizon, but
was also damaging those which spanned the boundary.

If affected by this bug, the symptoms vary depending on whether the table
is still live or if it is already dead:

- Live tables would have failures during any mutation request, when the history-tracking triggers attempt to access a storage table that is no longer present
- Live and dead tables vanish from catalog snapshots since no history storage is detected during snapshot access

The only way to recover from this damage is to restore the catalog from
an external DB backup.