wip: prevalidate and store "compiled" acl bindings
This uses plpgsql funcs to rewrite each constraint and column name in
an ACL binding into model RIDs and attach the model dependencies
needed for the policy invalidation trigger.  As a shim to the current
Python service logic, a decompilation step reverses this transform
during model introspection as ACL bindings are retrieved from DB.

This introduces a slight behavior change for weird cases.  If local
SQL has been used to rename a constraint or column while preserving
the associated model RIDs, the policy will be decompiled with the new
model names as well.  Previously, a stored policy would use the old
names; during introspection, such policies would have either been
discarded as invalid or resolved to different constraints/columns
having the old names.  This is not a scenario that could be produced
using just the REST APIs.

Remaining to do:
1. Perform compilation of ACL bindings during a catalog upgrade
2. Perform compilation of ACL bindings during history amendment