add test cases and fix for implicitly hiding keys when constituent columns are hidden
This was working for fkeys but not keys, and there were no regression
test cases for key visibility.

The bugfix is for the corner case where the key's constituent column
is non-enumerable but simultaneously has an ACL binding for "select"
rights. The internal ermrest rights model is a bit baroque in this
situation, and more specific logic tests were needed in the custom
"has_right()" method for the uniqueness constraint classes.

(cherry picked from commit d7527d17ae80ed78e87486925c35d9a5a696ce32)