fix internal error when resolving RIDs to a table that was dropped
There was a gap in the logic when a table was dropped with live
entities in it. Those may still be listed in the visible entities
tracking, but no longer have a known table entry for the desired
snaptime.

We have to look back and find the last visible snaptime, considering
all sources:

- visible_entities
- history tuples for resolved entity RID
- history tuples for table definition
- model and data snaptimes to find last catalog snaptime