simplify column alteration API by dropping the update query param
Since the service had to explicitly handle sparse inputs (with missing
fields), it seems simpler to just use this as the sole mechanism for
sparse updates.  The extra query parameter is confusingly redundant.

The only drawback to this approach is that a client must "opt-out" for
changes to individual fields rather than "opt-in", i.e. they would
prune elements from the input document rather than expressing a
positive action list in the URL query parameter. In an idiom where the
client obtains the current column document and selectively edits it,
this is not really an issue. The unedited fields will behave
idempotently and not cause any real change to the column definition
even if the client forgets to prune them.