Add support for PUT on table /key/cols resources to alter a key
These fields can be supplied to change the key:

- `names`: a replacement constraint-name
- `comment`: a replacment comment
- `annotations:` a replacement annotation map

The new constraint-name is obtained from `input['names'][0][1]`. If
the `names` list is empty or the first element fails to provide a
constraint name, the `names` field is ignored.  The supplied schema
name `input['names'][0][0]` SHOULD match the schema containing the
key's table, but MAY be ignored during request processing.

As with schema, table, and column alteration, omitting a field
means to make no change to that aspect of the model element.