extend the 503-based recovery to cover another inconsistency in acl bindings
We already have some logic to "prune" acl bindings if they are found
to be inconsistent during introspection. Previously, this was known to
occur if someone removed part of the model that was used by the acl
binding projection, i.e. a column/table/fkey.  These emerge as an
internal ConflictModel error during introspection of the binding.

Due to another bug, some other invalid policies might exist which
raise ConflictData. We can extend the recovery scheme to handle this
case as well, even though we should improve the invalidation to weed
out such errors earlier, during submission of the binding.

This patch also increases logging slightly in this error path, so that
other unexpected failures can more easily be investigated if they
occur in the wild. However, we will not blindly prune bindings for
such unrecognized errors. We should investigate first and understand
whether it is appropriate to extend the recovery logic on a
case-by-case basis.