add POST /entity/tablename?onconflict=skip for idempotent insert
This only supports the easiest scenario, but in a way that could
potentially be enhanced with more sub-options later. It uses the
existing "metakey" mechanism to distinguish colliding rows and simply
discards them.

This results in a first-writer-wins strategy for inserting rows only
if missing, but not attempting mix any update semantics into the
existing insert operation.