fix bug in POST /entity/ handling of default cols, surfaced by new onconflict=skip feature
A default column should not be part of the metakey, e.g. RID should
not be used to correlate input to existing rows if input isn't
supplying RIDs. This previously didn't really matter for POST /entity/
but matters now that we can ask ermrest to skip input rows which
collide.