add magic tag:isrd.isi.edu,2020:history-capture annotation for tables
Setting to JSON `false` will disable history capture. Setting to
JSON `true` is same as removing annotation and produces default history
behavior of ERMrest.  This hint is secondary to other ERMrest rules
and so is irrelevant if ERMrest has already decided to disable history,
i.e. due to absence of required system columns or because the table
kind is a view or other unusual relation in PostgresSQL.