expose alias_target concept via catalog and schema introspection
A catalog accessed as /ermrest/catalog/alias has the alias ID in the
URL. The "alias_target" field in the resource will expose the real
catalog ID, which might be important for a client wishing to emulate a
redirect or otherwise make reference to the real storage catalog.  For
example, to form permanent references to a snapshot of this catalog,
insulated from future changes to the alias binding which could route
to a totally different catalog with different history.