wip: working protoype of new /ermrest/alias/ resource space
Allows management of catalog aliases:

- POST /ermrest/alias/ <-- alias definition JSON
- GET /ermrest/alias/{id} --> alias definition + status JSON
- DELETE /ermrest/alias/{id}

This API works whether the alias is live or dangling. The aias
also appears as /ermrest/catalog/{id} if live.

The core alias definition is:

  {
    "id": "ALIAS",
    "owner": [ "admin role", ...],
    "alias_target": "TARGET"
  }

where TARGET must be an existing (real) catalog id. Recursive
aliases are disallowed, so each alias must point to an existing
storage catalog.