adjust ermrest-registry-purge to tolerate alias entries in registry
Normally, dead aliases should be deleted by the service, but if a DBA
intervenes or if this tool is run to purge ALL records, it can
encounter an alias record that it should delete. However, aliases do
not have a descriptor and so the tool should skip db-related cleanup
for those records.