allow custom btree indexing and slightly revise hueristics
For indexing preferences allow an optional "truthy" syntax:

   "btree": [ "colname", ... ]

to specify a custom column list for a btree index nominally managed
as the index for the column containing this hint.

Also, revise the default understanding of unique and fkey constraints
as an index hanging off the first column in the potentially composite
constraint.  An explicit or implied/default:

   "btree": true

will act much like "btree": [ "colname" ] as a single-column index.
Avoid redundant indexes for this default behavior, but allow custom
syntax above to introduce an index that might be redundant with
implicit unique indexes.