move timeout logic from data _GET() wrapper to ermpath.resource
We want to use timeouts to abort requests that run too long. Not just
for slow queries but for data mutation requests which might run long.
It is better to have postgres abort than to allow the web framework
to timeout and abandon transactions which continue to run on the DB.

The postgres primitive is a statement-level timeout, so we need to
calculate a new timeout incrementally before each potentially slow
statement that we execute.

The idea is to tell postgres to abort the statement if it exceeds the
remaining timeout budget for the request.  I.e. requests with multiple
statement executions must set a shorter timeout for each subsequent
statement.  We assume the initial request_timeout_s budget is
configured to be slightly shorter than the web framework timeouts so
we can hopefully raise an error to the client before the middleware
invents its own error.