bugfix for quantified RID value lists
The original `quantified_value_lists` feature has a bug where
`RID=any(...)` or `RID=all(...)` cause an internal server error,
logged as a NotImplementedError.

This patch corrects that and advertises a new `quantified_rid_lists`
feature flag to allow clients to detect servers where it is safe to use
this construct with the RID column.