change ETag to use hash of client's roles rather than just client name
This hash will change if the client changes group memberships,
to invalidate client-side caching of responses retrieved with the
prior permissions. This will help avoid stale or inconsistent results
most often seen when a user joins a new group to access additional
project-specific content and revisits a Chaise page accessed
recently with the old permissions.