fix to allow PUT ../column/X config updates when a column has a readonly type
An existing (legacy) column may have a "readonly" type that is not
permitted for new columns or as the target when altering a column
type.  But, other alterations not related to the type should still
be allowed.

The prior flaw was preventing even basic updates such as changing
column ACLs via the PUT ../column/X API used by deriva-py.