refactor column authz for consistent behavior of '*' projections
in order to provide a more consistent behavior between the /entity/
api and /attribute/ or /attributegroup/ variations with the '*'
special projection syntax, refactor column authz.  also update
related test cases.

these are the modes we want to better align:

- attribute projection by '*'
- aggregate projection by 'array(*)'
- filter predicates with '*' left-hand side

in these, the * indicates an expansion of columns similar
to the /entity/ API's implicit column list.  we want them to
have a similar logic around columns with enumerate=false
and/or select=false rights.

meanwhile, we want to preserve the different idiom:

- aggregate projection by 'cnt(*)'

which counts rows as in SQL, rather than expanding columns.