add is_persistent flag to registry
This will default TRUE and not really serve a purpose in typical
deployments.

A site might opt to deploy an auto-expiry script to delete
non-persistent catalogs after some lifetime is reached. They might
also revise the column default to FALSE and/or adjust policies
to decide which self-service clients can adjust the flag.