relax authz on GET /ermrest/catalog/N/history from owner to enumerate privilege
Similar to our RID resolution API, allow basic history range discovery
to any client with "enumerate" privilege on the whole catalog. E.g. this will
allow Chaise UI to discover the earliest available snapshot ID when
allowing a user to view historical content.

The test suite is updated to reflect that this request is allowed for all
our test clients, due to the catalog configuration used for these tests.